Simplifying GitHub class
